Stop dashboard client pages from blowing the 60s static budget
Wrap settings/home/billing/data/import in server pages that await
connection() so Cache Components does not SSR huge client trees during
build. Cap Vercel build workers to the machine core count to avoid
static-generation thrashing on 4-core builders.
