+++ app/[locale]/dashboard/dashboard-home-client.tsx
@@ -1,0 +1,177 @@
+"use client";
+
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { TradeTableReview } from "./components/tables/trade-table-review";
+import { AccountsOverview } from "./components/accounts/accounts-overview";
+import WidgetCanvas from "./components/widget-canvas";
+import { useEffect, useRef } from "react";
+import { useI18n } from "@/locales/client";
+import { useSearchParams } from "next/navigation";
+import { clearReferralCode } from "@/lib/referral-storage";
+
+export default function Home() {
+  const t = useI18n();
+  const mainRef = useRef<HTMLElement>(null);
+  const tabsListRef = useRef<HTMLDivElement>(null);
+  const searchParams = useSearchParams();
+
+  // Clear referral code after successful subscription
+  useEffect(() => {
+    const success = searchParams.get("success");
+    if (success === "true") {
+      // Clear referral code from localStorage after successful subscription
+      clearReferralCode();
+    }
+  }, [searchParams]);
+
+  useEffect(() => {
+    let timeoutId: NodeJS.Timeout;
+
+    const calculateHeight = () => {
+      // Clear any pending timeout
+      if (timeoutId) {
+        clearTimeout(timeoutId);
+      }
+
+      // Debounce the calculation to account for animations
+      timeoutId = setTimeout(() => {
+        // Get navbar height - it's sticky at the top
+        const navbar = document.querySelector(
+          'nav[class*="sticky"]'
+        ) as HTMLElement;
+        const navbarHeight = navbar?.offsetHeight || 96; // fallback to 96px
+
+        // Get tabs list height (wrapper includes py-2 padding)
+        const tabsList = tabsListRef.current;
+        const tabsListHeight = tabsList?.offsetHeight || 0;
+
+        // Calculate the actual bottom position of the fixed tabs wrapper
+        // Tabs wrapper is positioned at: navbarHeight from top
+        // Tabs wrapper bottom = navbarHeight + tabsListHeight
+        const tabsWrapperBottom = navbarHeight + tabsListHeight;
+
+        // Get main element's top position to calculate relative padding
+        const main = mainRef.current;
+        const mainTop = main?.getBoundingClientRect().top || 0;
+
+        // Calculate padding-top needed: tabs wrapper bottom - main top
+        // This ensures content starts right after the tabs wrapper
+        const calculatedPaddingTop = tabsWrapperBottom - mainTop;
+
+        // Set CSS custom property for navbar height
+        document.documentElement.style.setProperty(
+          "--navbar-height",
+          `${navbarHeight}px`
+        );
+        document.documentElement.style.setProperty(
+          "--tabs-height",
+          `${tabsListHeight}px`
+        );
+        document.documentElement.style.setProperty(
+          "--calculated-padding-top",
+          `${calculatedPaddingTop}px`
+        );
+      }, 100); // Small delay to account for animation
+    };
+
+    // Calculate on mount
+    calculateHeight();
+
+    // Recalculate on window resize
+    window.addEventListener("resize", calculateHeight);
+
+    // Create a ResizeObserver to watch for navbar height changes (when filters appear/disappear)
+    const resizeObserver = new ResizeObserver(calculateHeight);
+    const navbar = document.querySelector('nav[class*="sticky"]');
+    if (navbar) {
+      resizeObserver.observe(navbar);
+    }
+
+    // Create a MutationObserver to watch for DOM changes when ActiveFilterTags appear/disappear
+    const mutationObserver = new MutationObserver((mutations) => {
+      mutations.forEach((mutation) => {
+        // Check if filter tags were added or removed
+        if (mutation.type === "childList") {
+          const hasFilterTags =
+            Array.from(mutation.addedNodes).some(
+              (node) =>
+                node.nodeType === Node.ELEMENT_NODE &&
+                (node as Element).classList?.contains("border-t")
+            ) ||
+            Array.from(mutation.removedNodes).some(
+              (node) =>
+                node.nodeType === Node.ELEMENT_NODE &&
+                (node as Element).classList?.contains("border-t")
+            );
+
+          if (hasFilterTags) {
+            calculateHeight();
+          }
+        }
+      });
+    });
+
+    // Observe the navbar for DOM changes
+    if (navbar) {
+      mutationObserver.observe(navbar, {
+        childList: true,
+        subtree: true,
+      });
+    }
+
+    // Cleanup
+    return () => {
+      window.removeEventListener("resize", calculateHeight);
+      resizeObserver.disconnect();
+      mutationObserver.disconnect();
+      if (timeoutId) {
+        clearTimeout(timeoutId);
+      }
+    };
+  }, []);
+
+  return (
+    <main
+      ref={mainRef}
+      id="dashboard-content"
+      tabIndex={-1}
+      className="overflow-x-hidden"
+    >
+      <a
+        href="#dashboard-content"
+        className="sr-only focus:not-sr-only focus:fixed focus:left-4 focus:top-4 focus:z-50 focus:rounded-md focus:bg-background focus:px-4 focus:py-2 focus:text-sm focus:font-medium focus:text-foreground focus:shadow-md focus:ring-2 focus:ring-ring"
+      >
+        {t("dashboard.skipToContent")}
+      </a>
+      <Tabs defaultValue="widgets" className="w-full h-full pt-(--tabs-height,3rem)">
+        {/* Fixed TabsList positioned under navbar */}
+        <div
+          ref={tabsListRef}
+          className="fixed inset-x-0 top-(--navbar-height,5rem) z-30 w-full overflow-x-auto border-b bg-background shadow-xs"
+        >
+          <TabsList className="h-12 w-full min-w-max max-w-none rounded-none bg-muted/70 sm:min-w-0">
+            <TabsTrigger value="widgets">
+              {t("dashboard.tabs.widgets")}
+            </TabsTrigger>
+            <TabsTrigger value="table">{t("dashboard.tabs.table")}</TabsTrigger>
+            <TabsTrigger value="accounts">
+              {t("dashboard.tabs.accounts")}
+            </TabsTrigger>
+          </TabsList>
+        </div>
+
+        <TabsContent value="table" className="h-[calc(100dvh-var(--navbar-height,5rem)-var(--tabs-height,3rem)-16px)] min-w-0 p-2 sm:p-4">
+          <TradeTableReview />
+        </TabsContent>
+
+        <TabsContent value="accounts" className="h-[calc(100dvh-var(--navbar-height,5rem)-var(--tabs-height,3rem)-16px)] min-w-0 mt-0">
+          <AccountsOverview size="large" />
+        </TabsContent>
+
+        <TabsContent value="widgets" className="min-w-0 overflow-hidden px-0 max-md:h-[calc(100dvh-var(--navbar-height,5rem)-var(--tabs-height,3rem))] sm:px-4">
+          <WidgetCanvas />
+        </TabsContent>
+      </Tabs>
+    </main>
+  );
+}
